fix up types
